discs/requirements.txt
--find-links https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
transformers>=4.6.1
jax[cuda]
ml_collections
tensorflow
tensorflow_probability
matplotlib
absl-py
numpy
tqdm
clu
flax
optax
networkx
python-sat
tensorboard
pickle5
nltk